+++ digitalpy/core/zmanager/action_mapper.py
@@ -1,7 +1,7 @@
 from abc import ABC
-from digitalpy.routing.request import Request
+from digitalpy.core.zmanager.request import Request
 
-from digitalpy.routing.response import Response
+from digitalpy.core.zmanager.response import Response
 
 
 class ActionMapper(ABC):
