+++ digitalpy/core/zmanager/controller.py
@@ -1,9 +1,9 @@
 from abc import ABC
-from digitalpy.routing.action_mapper import ActionMapper
-from digitalpy.config.configuration import Configuration
+from digitalpy.core.zmanager.action_mapper import ActionMapper
+from digitalpy.core.digipy_configuration.configuration import Configuration
 from digitalpy.core.object_factory import ObjectFactory
-from digitalpy.routing.request import Request
-from digitalpy.routing.response import Response
+from digitalpy.core.zmanager.request import Request
+from digitalpy.core.zmanager.response import Response
 
 
 class Controller(ABC):
