+++ digitalpy/core/zmanager/response.py
@@ -1,5 +1,5 @@
 
-from digitalpy.routing.controller_message import ControllerMessage
+from digitalpy.core.zmanager.controller_message import ControllerMessage
 
 
 class Response(ControllerMessage):
